added declarations to tpu-client/polling-accounts-fetcher
