+++ package.json
@@ -4,27 +4,28 @@
 	"main": "lib/index.js",
 	"license": "MIT",
 	"dependencies": {
-		"@hubbleprotocol/hubble-sdk": "^1.0.21",
-		"@project-serum/anchor": "^0.22.0",
+		"@hubbleprotocol/hubble-sdk": "^1.0.23",
+		"@project-serum/anchor": "^0.22.1",
 		"@pythnetwork/client": "^2.6.0",
 		"@solana/spl-token": "^0.2.0",
-		"@solana/web3.js": "^1.35.1",
+		"@solana/web3.js": "^1.36.0",
 		"axios": "^0.26.0",
+		"bs58": "^5.0.0",
 		"decimal.js": "^10.3.1",
 		"dotenv": "^10.0.0",
-		"polling-account-fetcher": "^1.0.5",
-		"tpu-client": "^1.0.0"
+		"polling-account-fetcher": "^1.0.6",
+		"tpu-client": "^1.0.2"
 	},
 	"devDependencies": {
 		"@types/node": "^17.0.21",
-		"@typescript-eslint/eslint-plugin": "^4.28.0",
-		"@typescript-eslint/parser": "^4.28.0",
-		"eslint": "^7.29.0",
-		"eslint-config-prettier": "^8.3.0",
-		"eslint-plugin-prettier": "^3.4.0",
-		"prettier": "^2.4.1",
+		"@typescript-eslint/eslint-plugin": "^4.33.0",
+		"@typescript-eslint/parser": "^4.33.0",
+		"eslint": "^7.32.0",
+		"eslint-config-prettier": "^8.5.0",
+		"eslint-plugin-prettier": "^3.4.1",
+		"prettier": "^2.5.1",
 		"ts-node": "^10.6.0",
-		"typescript": "^4.5.5"
+		"typescript": "^4.6.2"
 	},
 	"scripts": {
 		"build": "ts-node build/index.ts",
