+++ tsconfig.json
@@ -10,6 +10,8 @@
 		"skipLibCheck": true,
 		"experimentalDecorators": true,
 		"emitDecoratorMetadata": true,
+		"declaration": true,
+		"declarationMap": true,
 		"baseUrl": ".",
 		"rootDir": "src",
 		"outDir": "./lib"
